Refactor tests and add new ones
