+++ tests/methods/schema/generate.test.js
@@ -1,0 +1,87 @@
+const { generateInputSchema } = require('../../../lib/index');
+const { fullSchema } = require('./schema');
+const { mockConsole, restoreConsole } = require('./console');
+
+describe('Generate input schemas', () => {
+	beforeEach(() => {
+		mockConsole();
+	});
+
+	afterEach(() => {
+		restoreConsole();
+	});
+
+	describe('Generate schemas', () => {
+		test('It should generate schema from requested key.', () => {
+			const generatedSchema = generateInputSchema({
+				schema: fullSchema,
+				keys: {
+					objectType: {},
+				},
+				operation: 'testRequestKey',
+			});
+			expect(generatedSchema).toEqual({
+				objectType: {
+					type: 'object',
+					title: 'Parent',
+					properties: {
+						name: {
+							type: 'string',
+						},
+						child: {
+							title: 'child',
+							properties: {
+								name: { type: 'string' },
+							},
+						},
+					},
+					additionalProperties: false,
+				},
+			});
+		});
+
+		test('It should generate schema with deep merge.', () => {
+			const generatedSchema = generateInputSchema({
+				schema: fullSchema,
+				keys: {
+					objectType: {
+						properties: {
+							child: {
+								description: 'Child description.',
+								properties: {
+									name: {
+										type: 'boolean',
+										default: true,
+									},
+								},
+							},
+						},
+					},
+				},
+				operation: 'testRequestKeyWithOverrides',
+			});
+			expect(generatedSchema).toEqual({
+				objectType: {
+					type: 'object',
+					title: 'Parent',
+					properties: {
+						name: {
+							type: 'string',
+						},
+						child: {
+							title: 'child',
+							description: 'Child description.',
+							properties: {
+								name: {
+									type: 'boolean',
+									default: true,
+								},
+							},
+						},
+					},
+					additionalProperties: false,
+				},
+			});
+		});
+	});
+});
